test: remove serve test
